One Way Streets
A demonstration of how to implement and make one-way connections in any direction without using the token `nowhere`.
(See Section 3.3 of the Writing with Inform for a style example using the predefined token `above`. Standard rules only define `above` and `below`.  We define additional tokens to handle directions other than up and down.)
1) Inform 7 story example
2) transcript (a complete walkthru of the example)